rename lifecycle.py to manager.py
