+++ src/vivarium/framework/lifecycle/manager.py
